[feat:#191][tsdb]: add suggesters for memdb and indexdb

+++ pkg/field/iterator.go
@@ -21,7 +21,7 @@ type GroupedTimeSeries interface {
 type MultiTimeSeries interface {
 	TimeSeries
 	// Version returns the version no.
-	Version() int64
+	Version() uint32
 	// ID returns the time series id under current metric
 	ID() uint32
 }
